Fixed detekt issues
