+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/SubmissionConfig.kt
@@ -5,6 +5,7 @@ import ac.uk.ebi.biostd.persistence.repositories.AccessPermissionRepository
 import ac.uk.ebi.biostd.persistence.repositories.TagsDataRepository
 import ac.uk.ebi.biostd.persistence.service.SubmissionRepository
 import ac.uk.ebi.biostd.submission.SubmissionSubmitter
+import ac.uk.ebi.biostd.submission.domain.service.ProjectService
 import ac.uk.ebi.biostd.submission.domain.service.SubmissionService
 import ac.uk.ebi.biostd.submission.domain.service.TempFileGenerator
 import ac.uk.ebi.biostd.submission.web.handlers.SubmissionWebHandler
@@ -27,12 +28,18 @@ class SubmissionConfig(
         subRepository: SubmissionRepository,
         serializationService: SerializationService,
         persistenceContext: PersistenceContext,
-        submissionSubmitter: SubmissionSubmitter,
+        submissionSubmitter: SubmissionSubmitter
+    ): SubmissionService =
+        SubmissionService(subRepository, persistenceContext, serializationService, submissionSubmitter)
+
+    @Bean
+    fun projectService(
+        submissionRepository: SubmissionRepository,
         accessPermissionRepository: AccessPermissionRepository,
         tagsDataRepository: TagsDataRepository
-    ): SubmissionService =
-        SubmissionService(subRepository, persistenceContext, serializationService, submissionSubmitter,
-            accessPermissionRepository, tagsDataRepository)
+    ): ProjectService =
+        ProjectService(submissionRepository, accessPermissionRepository, tagsDataRepository)
+
 
     @Bean
     fun submissionHandler(submissionService: SubmissionService): SubmissionWebHandler =
